tdf#131431 Move next node if pCurrentCursor hasn't moved in replaceAll
...to avoid an infinite loop

This commit addresses an issue where replacing an attribute at the
end of a paragraph with a new line results in an infinite loop.
The loop occurs because the cursor fails to advance
past the node boundary.

The root cause is an additional constraint that skips moving to the next node
when the cursor is at the beginning/end of the node. Introduced in commit
7087eb75b0e88429d5d2410c1aef54f30a86c560, the additional check aimed to
ensure that attributes in empty lines are not skipped.

This commit adds a check to detect when pCurrentCursor does not
advance and moves it forward.

Change-Id: I18ef6d38842ee4b06a72c9ae144eb69f2353cac9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185135
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins