tdf#139467 correctly handle DatabaseLocation for Calc cell ranges
svx::ODataAccessDescriptor contains properties for both a database
name and a local file database location. It parses a given name
via setDataSource() and puts the value into the correct property
depending on whether it starts with "file://" or not.

But this means that code which directly checks properties of
the ODataAccessDescriptor cannot just look for "DataSource"
but must also handle the "DatabaseLocation" property.

Net effect was that saving a Calc sheet with a cell range
backed by a database query from a local file database that was
NOT registered with LibreOffice would not save the database URL
to the .ods file, and on reload would lose the reference to the
database. This caused failures when trying to refresh the database
query.

On the UNO side creating import descriptors for a cell range
from a file with a database location would also fail to read
in the location. Fix that.

Change-Id: I7340f2d9f2f0552f436682f2b96249f39413c1fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195173
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Hossein <hossein@libreoffice.org>
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>