tdf#169651 Avoid crash in SwLayAction::FormatContent
Here is effort to avoid a crash that occurs when a fly frame is
unfloated that has a drawing object or fly frame anchored to its frame.

To avoid the crash, this approach uses FN_TOOL_ANCHOR_PARAGRAPH to
change the anchor of frame anchored drawing and fly frame objects to be
paragraph anchored in the fly frame content. MoveObjToVisibleLayer is
added in SwDrawContact::SwClientNotify
sw::DrawFrameFormatHintId::POST_RESTORE_FLY_ANCHOR case handling to make
undo restored fly anchored objects visible. SwFEShell::SelectFlyFrame is
made SW_DLLPUBLIC to make the clang linker not complain about 'linker
cannot find symbol(s) for architecture x86_64' when used in the included
unit test.

Change-Id: Ie6791cb45e7dce31132d3b6a8dac79d63bc27eb2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195650
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>